requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-brake==1.6.0
google-api-core==2.10.2
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-translate==2.0.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.48.2
idna==3.4
nose==1.3.7
oauthlib==3.2.2
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
